ohm: Add SO_REUSEPORT option to sockets
In its current form, I believe this breaks compatibility with
non-BSD unix OS's. This should be fairly easy to fix on a case by
case basis, particularly for linux  which should be just swapping
with SO_REUSEADDR.

Fixes a bug where the socket return value was incorrectly compared
to for the return value. This fixes the check (for -1) and accordingly
throws an esys log.